Explicitly instantiate uniform
