+++ source/framework/algorithm/include/lue/framework/algorithm/value_policies/uniform.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include "lue/framework/algorithm/uniform.hpp"
-#include "lue/concept.hpp"
 
 
 namespace lue {
@@ -10,7 +9,7 @@ namespace lue {
         // relevant. These have the same element type as the output element.
         template<Arithmetic Element, Arithmetic... SomeElement>
         using DefaultValuePolicies = policy::DefaultValuePolicies<
-            AllValuesWithinDomain<Element, Element>,
+            AllValuesWithinDomain<SomeElement..., Element, Element>,
             OutputElements<Element>,
             InputElements<SomeElement..., Element, Element>>;
 
@@ -23,8 +22,8 @@ namespace lue {
             requires(!std::is_same_v<Element, std::uint8_t> && !std::is_same_v<Element, std::int8_t>)
         auto uniform(
             PartitionedArray<SomeElement, rank> const& input_array,
-            hpx::shared_future<Element> const& min_value,
-            hpx::shared_future<Element> const& max_value) -> PartitionedArray<Element, rank>
+            Scalar<Element> const& min_value,
+            Scalar<Element> const& max_value) -> PartitionedArray<Element, rank>
         {
             using Policies = policy::uniform::DefaultValuePolicies<Element, SomeElement>;
 
@@ -39,18 +38,15 @@ namespace lue {
             Element const min_value,
             Element const max_value) -> PartitionedArray<Element, rank>
         {
-            return uniform(
-                input_array,
-                hpx::make_ready_future<Element>(min_value).share(),
-                hpx::make_ready_future<Element>(max_value).share());
+            using Policies = policy::uniform::DefaultValuePolicies<Element, SomeElement>;
+
+            return lue::uniform(Policies{}, input_array, min_value, max_value);
         }
 
 
         template<Arithmetic Element>
             requires(!std::is_same_v<Element, std::uint8_t> && !std::is_same_v<Element, std::int8_t>)
-        auto uniform(
-            hpx::shared_future<Element> const& min_value, hpx::shared_future<Element> const& max_value)
-            -> Scalar<Element>
+        auto uniform(Scalar<Element> const& min_value, Scalar<Element> const& max_value) -> Scalar<Element>
         {
             using Policies = policy::uniform::DefaultValuePolicies<Element>;
 
@@ -62,9 +58,9 @@ namespace lue {
             requires(!std::is_same_v<Element, std::uint8_t> && !std::is_same_v<Element, std::int8_t>)
         auto uniform(Element const min_value, Element const max_value) -> Scalar<Element>
         {
-            return uniform(
-                hpx::make_ready_future<Element>(min_value).share(),
-                hpx::make_ready_future<Element>(max_value).share());
+            using Policies = policy::uniform::DefaultValuePolicies<Element>;
+
+            return lue::uniform(Policies{}, min_value, max_value);
         }
