+++ source/framework/algorithm/source/miscellaneous_operation/uniform.cpp.in
@@ -1,0 +1,12 @@
+#include "lue/framework/algorithm/default_policies/uniform.hpp"
+#include "lue/framework/algorithm/definition/uniform.hpp"
+#include "lue/framework/algorithm/value_policies/uniform.hpp"
+
+
+namespace lue {
+
+    using Shape_ = Shape<Count, 2>;
+
+    LUE_INSTANTIATE_UNIFORM(ESC(policy::uniform::{{Policies}} < {{Element}} >), Shape_);
+
+}  // namespace lue
