+++ source/framework/api/cxx/CMakeLists.txt
@@ -90,6 +90,7 @@ target_link_libraries(lue_framework_api_cxx
     PRIVATE
         lue::framework_io
         lue::framework_local_operation
+        lue::framework_miscellaneous_operation
     PUBLIC
         lue::framework_algorithm
         lue::framework_partitioned_array
