+++ source/framework/python/CMakeLists.txt
@@ -56,6 +56,8 @@ add_library(lue_py_framework_local_operation SHARED
 add_library(lue_py_framework_miscellaneous_operation SHARED
     source/algorithm/miscellaneous_operation.cpp
     source/algorithm/miscellaneous_operation/resample.cpp
+    source/algorithm/miscellaneous_operation/uniform-1.cpp
+    source/algorithm/miscellaneous_operation/uniform-2.cpp
 )
 
 add_library(lue_py_framework_routing_operation SHARED
@@ -153,8 +155,6 @@ add_library(lue_py_framework SHARED
     $<$<BOOL:${LUE_FRAMEWORK_LOCALITY_ID_ELEMENT_AVAILABLE}>:source/algorithm/locality_id.cpp>
     $<$<BOOL:${LUE_FRAMEWORK_CLOCK_TICK_ELEMENT_AVAILABLE}>:source/algorithm/timestamp.cpp>
     source/algorithm/normal.cpp
-    source/algorithm/uniform-1.cpp
-    source/algorithm/uniform-2.cpp
 
     source/core/create_partitioned_array.cpp
     ${generated_source_files}
