+++ source/framework/python/source/algorithm/miscellaneous_operation/uniform-2.cpp
@@ -1,5 +1,6 @@
 #include "shape.hpp"
 #include "lue/framework/algorithm/value_policies/uniform.hpp"
+#include "lue/framework/core/domain_decomposition.hpp"
 #include "lue/framework.hpp"
 #include <pybind11/numpy.h>
 #include <pybind11/stl.h>
@@ -23,11 +24,12 @@ namespace lue::framework {
 
             if constexpr (arithmetic_element_supported<Element>)
             {
-                result = pybind11::cast(value_policies::uniform(
-                    array_shape,
-                    partition_shape,
-                    pybind11::cast<Element>(min_value),
-                    pybind11::cast<Element>(max_value)));
+                result = pybind11::cast(
+                    value_policies::uniform(
+                        array_shape,
+                        partition_shape,
+                        pybind11::cast<Element>(min_value),
+                        pybind11::cast<Element>(max_value)));
             }
 
             return result;
@@ -141,10 +143,11 @@ namespace lue::framework {
 
                 if (dynamic_array_shape.size() != dynamic_partition_shape.size())
                 {
-                    throw std::runtime_error(std::format(
-                        "Rank of array shape and partition shape must be equal ({} != {})",
-                        dynamic_array_shape.size(),
-                        dynamic_partition_shape.size()));
+                    throw std::runtime_error(
+                        std::format(
+                            "Rank of array shape and partition shape must be equal ({} != {})",
+                            dynamic_array_shape.size(),
+                            dynamic_partition_shape.size()));
                 }
             }
